requirements-dev.txt
coverage==5.2.1
pre-commit
pylint
pylint-flask
black
pdbpp
